move Drive initialization to store.ts, which is already doing program init stuff. Editor doesn't need this constructor and can be driven by state updates
